feat: move credential status indicator to app header
Move the credential status from the settings page into the app header
(mobile) and sidebar footer (desktop) so it's always visible. The
indicator shows a colored dot (green/yellow/red) with detailed status
in a tooltip on hover.

Closes #88
